Fixed compiler warnings
Mostly -Wsign, extra ;, unused functions reported by -Wall -Wextra -Wpedantic

+++ include/cgimap/options.hpp
@@ -13,26 +13,26 @@ class global_settings_base {
 public:
   virtual ~global_settings_base();
 
-  virtual long get_payload_max_size() const = 0;
-  virtual int get_map_max_nodes() const = 0;
+  virtual uint32_t get_payload_max_size() const = 0;
+  virtual uint32_t get_map_max_nodes() const = 0;
   virtual double get_map_area_max() const = 0;
   virtual std::string get_changeset_timeout_open_max() const = 0;
   virtual std::string get_changeset_timeout_idle() const = 0;
-  virtual int get_changeset_max_elements() const = 0;
-  virtual int get_way_max_nodes() const = 0;
-  virtual long get_scale() const = 0;
-  virtual boost::optional<int> get_relation_max_members() const = 0;
-  virtual boost::optional<int> get_element_max_tags() const = 0;
+  virtual uint32_t get_changeset_max_elements() const = 0;
+  virtual uint32_t get_way_max_nodes() const = 0;
+  virtual int64_t get_scale() const = 0;
+  virtual boost::optional<uint32_t> get_relation_max_members() const = 0;
+  virtual boost::optional<uint32_t> get_element_max_tags() const = 0;
 };
 
 class global_settings_default : public global_settings_base {
 
 public:
-  long get_payload_max_size() const override {
+  uint32_t get_payload_max_size() const override {
     return 50000000L;
   }
 
-  int get_map_max_nodes() const override {
+  uint32_t get_map_max_nodes() const override {
     return 50000;
   }
 
@@ -48,23 +48,23 @@ public:
      return "1 hour";
   }
 
-  int get_changeset_max_elements() const override {
+  uint32_t get_changeset_max_elements() const override {
      return 10000;
   }
 
-  int get_way_max_nodes() const override {
+  uint32_t get_way_max_nodes() const override {
      return 2000;
   }
 
-  long get_scale() const override {
+  int64_t get_scale() const override {
      return 10000000L;
   }
 
-  boost::optional<int> get_relation_max_members() const override {
+  boost::optional<uint32_t> get_relation_max_members() const override {
      return boost::none;  // default: unlimited
   }
 
-  boost::optional<int> get_element_max_tags() const override {
+  boost::optional<uint32_t> get_element_max_tags() const override {
      return boost::none;  // default: unlimited
   }
 };
@@ -87,11 +87,11 @@ public:
     set_new_options(options);
   }
 
-  long get_payload_max_size() const override {
+  uint32_t get_payload_max_size() const override {
     return m_payload_max_size;
   }
 
-  int get_map_max_nodes() const override {
+  uint32_t get_map_max_nodes() const override {
     return m_map_max_nodes;
   }
 
@@ -107,23 +107,23 @@ public:
      return m_changeset_timeout_idle;
   }
 
-  int get_changeset_max_elements() const override {
+  uint32_t get_changeset_max_elements() const override {
      return m_changeset_max_elements;
   }
 
-  int get_way_max_nodes() const override {
+  uint32_t get_way_max_nodes() const override {
      return m_way_max_nodes;
   }
 
-  long get_scale() const override {
+  int64_t get_scale() const override {
      return m_scale;
   }
 
-  boost::optional<int> get_relation_max_members() const override {
+  boost::optional<uint32_t> get_relation_max_members() const override {
      return m_relation_max_members;
   }
 
-  boost::optional<int> get_element_max_tags() const override {
+  boost::optional<uint32_t> get_element_max_tags() const override {
      return m_element_max_tags;
   }
 
@@ -142,16 +142,16 @@ private:
   void set_element_max_tags(const po::variables_map &options);
   bool validate_timeout(const std::string &timeout) const;
 
-  long m_payload_max_size;
-  int  m_map_max_nodes;
+  uint32_t m_payload_max_size;
+  uint32_t  m_map_max_nodes;
   double m_map_area_max;
   std::string m_changeset_timeout_open_max;
   std::string m_changeset_timeout_idle;
-  int m_changeset_max_elements;
-  int m_way_max_nodes;
-  long m_scale;
-  boost::optional<int> m_relation_max_members;
-  boost::optional<int> m_element_max_tags;
+  uint32_t m_changeset_max_elements;
+  uint32_t m_way_max_nodes;
+  int64_t m_scale;
+  boost::optional<uint32_t> m_relation_max_members;
+  boost::optional<uint32_t> m_element_max_tags;
 };
 
 class global_settings final {
@@ -162,10 +162,10 @@ public:
   static void set_configuration(std::unique_ptr<global_settings_base> && b) { settings = std::move(b); }
 
   // Maximum Size of HTTP body payload accepted by uploads, after decompression
-  static long get_payload_max_size() { return settings->get_payload_max_size(); }
+  static uint32_t get_payload_max_size() { return settings->get_payload_max_size(); }
 
   // Maximum number of nodes returned by the /map endpoint
-  static int get_map_max_nodes() { return settings->get_map_max_nodes(); }
+  static uint32_t get_map_max_nodes() { return settings->get_map_max_nodes(); }
 
   // Maximum permitted area for /map endpoint
   static double get_map_area_max() { return settings->get_map_area_max(); }
@@ -177,19 +177,19 @@ public:
   static std::string get_changeset_timeout_idle() { return settings->get_changeset_timeout_idle(); }
 
   // Maximum number of elements permitted in one changeset
-  static int get_changeset_max_elements() { return settings->get_changeset_max_elements(); }
+  static uint32_t get_changeset_max_elements() { return settings->get_changeset_max_elements(); }
 
   // Maximum number of nodes permitted in a way
-  static int get_way_max_nodes() { return settings->get_way_max_nodes(); }
+  static uint32_t get_way_max_nodes() { return settings->get_way_max_nodes(); }
 
   // Conversion factor from double lat/lon format to internal int format used for db persistence
-  static long get_scale() { return settings->get_scale(); }
+  static int64_t get_scale() { return settings->get_scale(); }
 
   // Maximum number of relation members for an OSM object (may be unlimited)
-  static boost::optional<int> get_relation_max_members() { return settings->get_relation_max_members(); }
+  static boost::optional<uint32_t> get_relation_max_members() { return settings->get_relation_max_members(); }
 
   // Maximum number of tags for an OSM object (may be unlimited)
-  static boost::optional<int> get_element_max_tags() { return settings->get_element_max_tags(); }
+  static boost::optional<uint32_t> get_element_max_tags() { return settings->get_element_max_tags(); }
 
 private:
   static std::unique_ptr<global_settings_base> settings;  // gets initialized with global_settings_default instance
